social_graph/requirements.txt
astroid==2.12.13
click==8.1.3
dill==0.3.6
Flask==2.2.2
isort==5.11.4
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
lazy-object-proxy==1.8.0
MarkupSafe==2.1.1
mccabe==0.7.0
mysql-connector-python==8.0.31
parso==0.8.3
platformdirs==2.6.2
protobuf==3.20.1
pylint==2.15.9
tomli==2.0.1
tomlkit==0.11.6
Werkzeug==2.2.2
wrapt==1.14.1
